requirements.txt
streamlit
pandas
Pillow
pytesseract
openai
python-dotenv
